feat(predictor): add counselling options UI and context setup
Introduce a new counselling options UI with responsive grid layout,
hover and active states for better interactivity and visual feedback.
Add CategoryContext as a placeholder for future category state management.
Refactor predictor page styles to use CSS modules for scoped styling.
These changes improve UI clarity and prepare context for state sharing.
