+++ src/app/predictor/[counselling]/styles.module.css
