+++ src/context-providers/category-context-provider.tsx
@@ -1,0 +1,7 @@
+export default function CategoryContextProvider() {
+    return (
+        <div>
+            {/* Your context provider logic goes here */}
+        </div>
+    )
+}
