+++ src/contexts/category-context.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+export const CategoryContext = createContext<{
+
+}[]>([])
